requirements.txt
autogluon==1.4.0
redis==5.1.0
psycopg2-binary==2.9.9
requests==2.32.3
celery==5.4.0
python-dotenv
xai-sdk
holidays
pytz
yfinance
pandas
numpy
